Impl. add task to list
Created ChildItem task to represent the tasks in the lists.
Created AddTaskActivity to handle creation of a new task to a specific list.
Updated ExpandableListAdapter to handle a list of ChildItems instead of Strings
Implemented the functionality in MainActivity to start the new activity for adding a task when dobbelt clicking on a list, and to save the task in memory